pyproject.toml
[project]
name = "p2i-gan-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8,<3.9"
dependencies = [
    "decord>=0.6.0",
    "h5py>=3.11.0",
    "imageio>=2.35.1",
    "mlflow>=2.17.2",
    "pyyaml>=6.0.0",
    "torch>=2.4.1",
    "torchaudio>=2.4.1",
    "torchmetrics>=1.0.3",
    "torchvision>=0.19.1",
    "tqdm>=4.66.5",
    "zarr>=2.16.1",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["p2igan_bench*", "experiments*"]
